requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38h2bbff1b_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py38haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.4=py38h080aedc_0
ca-certificates=2022.5.18.1=h5b45459_0
certifi=2022.5.18.1=py38haa244fe_0
cffi=1.15.0=py38h2bbff1b_1
click=8.0.4=py38haa95532_0
colorama=0.4.4=pyhd3eb1b0_0
console_shortcut=0.1.1=4
cycler=0.11.0=pyhd8ed1ab_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.5.1=py38hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py38haa95532_0
et_xmlfile=1.0.1=py_1001
executing=0.8.3=pyhd3eb1b0_0
flask=2.0.3=pyhd3eb1b0_0
freeglut=3.2.2=h0e60522_1
freetype=2.10.4=h546665d_1
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=69.1=h0e60522_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2022.1.0=h57928b3_3787
ipykernel=6.9.1=py38haa95532_0
ipython=8.3.0=py38haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
itsdangerous=2.0.1=pyhd3eb1b0_0
jasper=2.0.33=h77af90b_0
jedi=0.18.1=py38haa95532_1
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h8ffe710_1
jsonschema=4.4.0=py38haa95532_0
jupyter_client=7.2.2=py38haa95532_0
jupyter_core=4.10.0=py38haa95532_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.4.2=py38hbd9d945_1
lerc=3.0=h0e60522_0
libclang=13.0.1=default_h81446c8_0
libdeflate=1.10=h8ffe710_0
libopencv=4.0.1=hbb9e17c_0
libpng=1.6.37=h1d00b33_2
libprotobuf=3.20.1=h7755175_0
libtiff=4.3.0=hc4061b1_4
libwebp-base=1.2.2=h8ffe710_1
libzlib=1.2.12=h8ffe710_0
lz4-c=1.9.3=h8ffe710_1
mahotas=1.4.11=py38h5d928e2_2
markupsafe=2.0.1=py38h2bbff1b_0
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py38hbe5506b_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py38he774522_1000
mkl=2019.4=245
mkl-service=2.3.0=py38h196d8e1_0
mkl_fft=1.3.0=py38h46781fe_0
mkl_random=1.1.0=py38hf9181ef_0
mysql-connector-python=8.0.29=pypi_0
nbclient=0.5.13=py38haa95532_0
nbconvert=6.4.4=py38haa95532_0
nbformat=5.3.0=py38haa95532_0
nest-asyncio=1.5.5=py38haa95532_0
notebook=6.4.11=py38haa95532_0
numexpr=2.7.3=py38hcbcaa1e_0
numpy=1.19.2=py38hadc3359_0
numpy-base=1.19.2=py38ha3acd2a_0
opencv=4.0.1=py38h2a7c758_0
openpyxl=3.0.9=pyhd8ed1ab_0
openssl=1.1.1o=h8ffe710_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py38hd77b12b_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pip=21.2.2=py38haa95532_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
pure_eval=0.2.2=pyhd3eb1b0_0
py-opencv=4.0.1=py38he44ac1e_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.12.3=py38h7ae7562_4
pyqt5-sip=4.19.18=pypi_0
pyqtchart=5.12=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyrsistent=0.18.0=py38h196d8e1_0
python=3.8.13=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytz=2021.3=pyhd3eb1b0_0
pywin32=302=py38h2bbff1b_2
pywinpty=2.0.2=py38h5da7b33_0
pyzmq=22.3.0=py38hd77b12b_2
qt=5.12.9=h556501e_6
scikit-learn=1.0.2=py38hf11a4ad_1
scipy=1.6.2=py38h14eb087_0
seaborn=0.11.2=pyhd3eb1b0_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=61.2.0=py38haa95532_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.38.3=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
tbb=2021.5.0=h2d74725_1
terminado=0.13.1=py38haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tornado=6.1=py38h2bbff1b_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py38haa95532_2
winpty=0.4.3=4
xz=5.2.5=h62dcd97_1
zipp=3.8.0=py38haa95532_0
zlib=1.2.12=h8ffe710_0
zstd=1.5.2=h6255e5f_1
